Poprawka 1 z dnia 20.06.17
Poprawia małe niedopatrzenie w kodzie.

+++ Katalogowanie plikow/PRI-KATALOGOWANIE-PLIKÓW/Karol_main.cs
@@ -273,7 +273,7 @@ namespace PRI_KATALOGOWANIE_PLIKÓW
             form2.Owner = this;
             form2.ShowDialog();
             proceed = form2.proceed;
-            System.IO.File.Delete(program_path + @"\temp\thumbnail.png");
+            if(File.Exists(program_path + @"\temp\thumbnail.png")) System.IO.File.Delete(program_path + @"\temp\thumbnail.png");
             extractionOptions.DEBUG_displayTimeRanges();
             if (proceed)
             {
